pagent-os/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
